requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
youtube-transcript-api
python-dotenv
